Update migration logic
